+++ rusoto/credential/src/claims.rs
@@ -14,16 +14,16 @@ pub const SUBJECT: &'static str = "sub";
 /// Key used in the claims map for the `Audience` claim.
 ///
 /// The intended audience (also known as client ID) of the web identity token.
-/// This is traditionally the client identifier issued to the application that requested the web identity token. 
+/// This is traditionally the client identifier issued to the application that requested the web identity token.
 /// For OpenID Connect this field contains the value of the `aud` claim.
 /// For SAML this is the value of the `Recipient` attribute of the `SubjectConfirmationData`
-/// element of the SAML assertion. 
+/// element of the SAML assertion.
 pub const AUDIENCE: &'static str = "aud";
 
 /// Key used in the claims map for the `Issuer` claim.
 ///
-/// For OpenID Connect ID Tokens this contains the value of the `iss` field. 
-/// For OAuth 2.0 access tokens, this contains the value of the `ProviderId` parameter 
-/// that was passed in the `AssumeRoleWithWebIdentity` request. 
-/// For SAML this is the value of the Issuer element of the SAML assertion. 
+/// For OpenID Connect ID Tokens this contains the value of the `iss` field.
+/// For OAuth 2.0 access tokens, this contains the value of the `ProviderId` parameter
+/// that was passed in the `AssumeRoleWithWebIdentity` request.
+/// For SAML this is the value of the Issuer element of the SAML assertion.
 pub const ISSUER: &'static str = "iss";