+++ rusoto/credential/src/container.rs
@@ -1,10 +1,12 @@
 //! The Credentials provider to read from a task's IAM Role.
 
-use reqwest;
+use hyper::Uri;
 use std::env::var as env_var;
-use std::io::Read;
+use std::time::Duration as StdDuration;
+use tokio_core::reactor::Core;
 
-use {AwsCredentials, CredentialsError, ProvideAwsCredentials, parse_credentials_from_aws_service};
+use {AwsCredentials, CredentialsError, ProvideAwsCredentials, ProvideTimeoutableAwsCredentials,
+     make_request, parse_credentials_from_aws_service};
 
 const AWS_CREDENTIALS_PROVIDER_IP: &'static str = "169.254.170.2";
 
@@ -14,12 +16,23 @@ pub struct ContainerProvider;
 
 impl ProvideAwsCredentials for ContainerProvider {
     fn credentials(&self) -> Result<AwsCredentials, CredentialsError> {
-        credentials_from_container()
+        credentials_from_container(None)
+    }
+}
+
+impl ProvideTimeoutableAwsCredentials for ContainerProvider {
+    fn credentials_with_timeout(
+        &self,
+        timeout: StdDuration,
+    ) -> Result<AwsCredentials, CredentialsError> {
+        credentials_from_container(Some(timeout))
     }
 }
 
 /// Grabs the Credentials from the AWS Container Credentials Provider. (169.254.170.2).
-fn credentials_from_container() -> Result<AwsCredentials, CredentialsError> {
+fn credentials_from_container(
+    timeout: Option<StdDuration>,
+) -> Result<AwsCredentials, CredentialsError> {
     let aws_container_credentials_relative_uri =
         match env_var("AWS_CONTAINER_CREDENTIALS_RELATIVE_URI") {
             Ok(v) => v,
@@ -30,29 +43,14 @@ fn credentials_from_container() -> Result<AwsCredentials, CredentialsError> {
             }
         };
     let address: String = format!(
-        "http://{}{}",
+        "http://{}{}/",
         AWS_CREDENTIALS_PROVIDER_IP,
         aws_container_credentials_relative_uri
     );
-    let response = reqwest::get(&address);
-
-    match response {
-        Ok(mut resp) => {
-            if resp.status().is_success() {
-                let mut body = String::new();
-                if resp.read_to_string(&mut body).is_err() {
-                    return Err(CredentialsError::new(
-                        "Didn't get a parsable response body from the credentials provider",
-                    ));
-                }
-                parse_credentials_from_aws_service(&body)
-            } else {
-                return Err(CredentialsError::new(format!(
-                    "Couldn't connect to credentials provider: {}",
-                    resp.status()
-                )));
-            }
-        }
+    let core = try!(Core::new());
+    let uri = try!(address.parse::<Uri>());
+    match make_request(core, uri, timeout) {
+        Ok(resp) => parse_credentials_from_aws_service(&resp),
         Err(err) => {
             return Err(CredentialsError::new(
                 format!("Couldn't connect to credentials provider: {}", err),